+++ siteapp/urls.py
@@ -87,7 +87,7 @@ urlpatterns = [
     # must be last because regex matches some previous URLs
     url(r'^projects/(\d+)/(?:[\w\-]+)(/list)$', views.project_list_all_answers),
     # must be last because regex matches some previous URLs
-    url(r'^projects/(\d+)/(?:[\w\-]+)(/outputs)$', views.project_outputs),
+    url(r'^projects/(\d+)/(?:[\w\-]+)(/artifacts)$', views.project, name="project_artifacts"),
     # must be last because regex matches some previous URLs
     url(r'^projects/(\d+)/(?:[\w\-]+)(/api)$', views.project_api),
     # must be last because regex matches some previous URLs