+++ api/base/views/mixins.py
@@ -71,7 +71,7 @@ class CustomSearchModelMixin(object):
         pass
 
     def _search(self, request):
-        keyword = request.GET.get('keyword')
+        keyword = request.GET.get('search')
         if keyword:
             return self.search(request, keyword)