+++ api/controls/serializers/controls.py
@@ -1,7 +1,8 @@
 from api.base.serializers.types import ReadOnlySerializer
 from api.controls.serializers.element import DetailedElementSerializer
 from controls.models import CommonControlProvider, CommonControl, ElementCommonControl
-
+from controls.oscal import CatalogData
+from rest_framework import serializers
 
 class SimpleCommonControlProviderSerializer(ReadOnlySerializer):
     class Meta:
@@ -36,3 +37,29 @@ class DetailedElementCommonControlSerializer(SimpleElementCommonControlSerialize
     class Meta:
         model = ElementCommonControl
         fields = SimpleElementCommonControlSerializer.Meta.fields + ['element', 'common_control']
+
+
+class CatalogDataReadSerializer(ReadOnlySerializer):
+
+    title = serializers.SerializerMethodField()
+
+    def get_title(self, obj):
+        return obj.catalog_json['catalog']['metadata']['title'].strip()
+
+    class Meta:
+        model = CatalogData
+        fields = ['catalog_key', 'title']
+
+class CatalogDataGroup(serializers.Serializer):
+    id = serializers.CharField()
+    title = serializers.CharField()
+
+class CatalogDataWithGroupReadSerializer(CatalogDataReadSerializer):
+    groups = serializers.SerializerMethodField()
+
+    def get_groups(self, obj):
+        return [CatalogDataGroup(x).data for x in obj.catalog_json['catalog'].get('groups', [])]
+
+    class Meta:
+        model = CatalogData
+        fields = CatalogDataReadSerializer.Meta.fields + ['groups']