+++ CHANGELOG.md
@@ -6,6 +6,13 @@ v0.9.13-dev (January xx, 2022)
 ------------------------------
 
 
+**UI changes**
+
+* Add button to mark component as deleted.
+
+**Developer changes**
+
+* Add support for soft delete of components/elements by adding `deleted` field to `controls.Elements` model.
 
 v0.9.11.11 (January 15, 2022)
 -----------------------------