+++ CHANGELOG.md
@@ -5,7 +5,6 @@ GovReady-Q Release Notes
 v0.9.13-dev (January xx, 2022)
 ------------------------------
 
-
 **UI changes**
 
 * Add button to mark component as deleted.
@@ -13,6 +12,11 @@ v0.9.13-dev (January xx, 2022)
 **Developer changes**
 
 * Add support for soft delete of components/elements by adding `deleted` field to `controls.Elements` model.
+
+**Bug fixes**
+
+* Properly close CSV export modal after exporting.
+
 
 v0.9.11.11 (January 15, 2022)
 -----------------------------